Remove many unneeded flags & simplify eval_proc_main
Eval_proc_main required that many flags set during codegen are passed when simulating a block. Most of the information these flags provide are already present in the block signature proto so use that instead.

Also change most test IRs to be based on a DSLX implementation. The eval_proc_main is only designed to evaluate the ir for designs compiled with a current toolchain so freezing IR only allows definitions to drift. Due to relying on the exact naming of nodes the observer test could not be changed.

Add a few extra tests for corner-cases not covered by previous tests.

PiperOrigin-RevId: 681946172