[xls] Simplify arith_to_xls libcall patterns
Generate call_dslx ops instead of vectorized_call ops. Teach the scalarizer how to
scalarize call_dslx ops.

Removes the getLibcall...() function that was hairy plus a bunch of code working out
what the libcall return type should be (now it's just picked up by the CallDslxOp
because the pattern is trivial enough that it can be inferred).

The same can be done to the math patterns; I'll do that next.

PiperOrigin-RevId: 681969450