Add strictness property to ChannelReferences. ChannelReferences are used in procs with proc-scoped channels (new-style procs) for the interface of procs and within procs to refer to proc-scoped channels. ChannelReferences are bound to channels in elaboration. The issue is that there was no way to specify the channel strictness of the interface of a proc because these are defined by channel references.
With this change, the interface channels of new style procs now has keyword arguments and look like:

  proc my_proc<in_ch: bits[32] in kind=single_value,
               out_ch: bits[32] out kind=streaming
                  strictness=proven_mutually_exclusive>(

It likely is better to have ChannelReferences have no properties (no strictness or kind) and instead have a construct which takes the place of a channel in the interface of a proc (channel stub maybe?). This construct can hold channel-like properties. In this world ChannelReferences are dumb references to channels. I may revisit this after finishing the proc-scoped channel refactor when all such issues have been exposed.

PiperOrigin-RevId: 682116348