[xls][mlir] Fix use of outlineSingleBlockRegion for EprocOps
LowerCountedFor uses outlineSingleBlockRegion as a helper, but inside it is hardcoded
to look for a parent FuncOp and dereferences it even if it is null. This causes problems
for ForOps inside EprocOps because those, obviously, are not FuncOps.

As an expedient we have copied the function with the one-line change to look
for a SymbolOpInterface instead, but this should be upstreamed.

PiperOrigin-RevId: 684409026