Restore the ability to incrementally add nominal type dims.
In implementing the fix for #1615, we added `ResolveNominalTypeDims` function, which is used by `ResolveInternal` as a safer implementation of its use case than pushing a map of possibly extraneous dims through `AddNominalTypeDims`.

It was theorized that we then no longer needed the ability for `ConcretizeStructAnnotation` to incrementally populate dims, but this was  disproven by code in the zstd_dec_verilog target on a branch. This change restores the incremental add support, but keeps it only used by `ConretizeStructAnnotation`.

Fixes #1655

PiperOrigin-RevId: 684489422