[XLS] Improve recovery after minimizing clock
Fix one bug that let explained errors prevent recovery; if the user asked us to try to recover with the shortest viable clock, we should do so even if the failure was explainable.

Also, when recovering, apply `period_relaxation_percent` to the clock found, just as we would in a normal case where the clock was unspecified.

PiperOrigin-RevId: 685798413