[xls] Fix use-after-free error in NameCollisionResolver
Found by asan. names_.update(new_name) takes the internal storage of `new_name`,
which is a local variable that happens to be move-returned to the caller. Once
the caller has done with this variable the storage is destroyed.

This code is a little awkward to fix; a natural fix might be to use the storage
inside name_updates_ but of course that is invalidated on rehash. We could do
clever things like using a node_hash_map, but I took the expedient approach of
explicitly storing the string using a std::deque (deque does not invalidate iterators
on expand unlike std::vector).

PiperOrigin-RevId: 686491297