[DSLX:LS] Diagnostics updates for dependent modules.
Previously if you made a change to `leaf.x` that unbroke a module that
depended on `leaf.x`, you wouldn't naturally get an update to the
"problems" pane in VSCode that indicated the problem had been resolved.

Introduces a very lightweight virtualized filesystem layer so we can see
updated text in the editor (language server) workspace reflected in the
language server results.

Update before review.