[XLScc] Use attributes for all annotations
This CL replaces the use of intrinsics for (e.g.) loop unrolling and pipelining with custom statement attributes; it also finishes implementing inline channel strictness specifications as an attribute. This lets clang do all of the heavy lifting in terms of associating the annotation with its subject.

We continue to support the pragma-based forms for compatibility with other C++-based HLS implementations, but these are internally translated into the corresponding attributes.

PiperOrigin-RevId: 689002884