Add a query-engine wrapper to track changes to a graph.
It is often useful to use a query engine as one is modifying the graph by adding new nodes (eg for narrowing). When one does this it can cause problems however by making the arguments to individual nodes different complicating analysis of those nodes. To fix this a we add a wrapper for query engines that one can use to alias nodes to other nodes analysis results. This means that if one replaces eg `(+ a b)` with `(zero-extend (+ (bitslice a) (bitslice b)))` one can ensure that any users that request intervals or ranges of the new 'zero-extend' will see the information for the original '+' node.

Integrate this into the narrowing pass.

PiperOrigin-RevId: 689500577