[xls][mlir] Add a min_pipeline_stages attribute to EprocOp and SprocOp
This defaults to 1, and allows the compiler to indicate the minimum number of pipeline stages
that are *required* for the program to compile. This does not guarantee that the program
will compile in exactly these stages, but more an assertion that the program is impossible
to compile in fewer than min_pipeline_stages.

Increment min_pipeline_stages in convertForOpToSprocCall(), as we add a recv(send()) sequence.

PiperOrigin-RevId: 689750193