Make node functions for input nodes.
`Param`, `InputPort`, and `RegisterRead` nodes previously had no node function implementation in the JIT because they are implemented simply by having a pointer to memory. However, this meant these nodes needed special handling in the JIT for observers as they had no node function to invoke the required callbacks. This is problematic for a future refactoring of the JIT that will remove references to the `Package` (or not construct a package for AOT) because the JIT uses the `Package` to find `Node*` values for these input nodes.

This change adds node function impelementations for these nodes. These functions have no outputs- the previous handling of inputs is kept the same. These functions are purely to invoke callbacks. This change allows the JIT runtimes to remove some special-case handling of observers.

PiperOrigin-RevId: 691469959