Refactor JIT to reduce reliance on `Package`.
The various JIT "runtimes" (this term is super overloaded, but context hopefully makes what I mean clear) rely on some metadata from the function/proc/block, even after object code is generated. To use the object code, you need to know the number of inputs, their types, and their names. The JIT currently keeps a reference to the corresponding `Function*/Proc*/Block*` to get this metadata when needed.

This approach makes the AOT flow a little tricky as there isn't a pre-existing package to hold the metadata. The current state of affairs is that we serialize the IR to generated source and parse it when constructing a wrapper for the AOT code. This approach has some downsides:

- This parsing causes scalability issues. Parsing this IR can be slow for large designs. Even if the compiled IR is small, there might be a lot of other stuff in the package that makes constructing an AOT wrapper slow.
- This approach can lead to subtle bugs. The block JIT performs inlining in multiple places, but you could accidentally use un-inlined IR to poke at the JIT. This can result in some internal expectation being violated, for example  it may try to get the type of a register that doesn't exist in the un-inlined IR.

Fortunately, the JITs don't need very much information from the `Package` so it's relatively straightforward to summarize that information without relying on a pre-existing `Package`. The required information is:

1) `Function/Proc/Block` name
2) Input/output names
3) Input/output types

The JIT uses its own internal `TypeManager` to hold types, but the types are never exposed externally. They are only used to check that `Value`s are compatible with their destination and to marshall data to/from the JIT.

The runtime JIT populates this metadata directly from the package. The AOT flow instead uses the `AotEntrypointProto` which now includes a `PackageInterfaceProto` variant for the compiled `Function/Proc/Block`. This interface proto is sufficient to populate the metadata without the expense of parsing all the IR. This significantly speeds up AOT JIT wrapper construction for large designs.

Note that what I've described above is true for `Function` and `Block` JITs, but the `Proc` JIT is different because of how the runtime manages channels and inter-proc communication. The old approach is largely in place for `Proc`s.

PiperOrigin-RevId: 691470815