Store sample inputs as testvector::SampleInputsProto
There are various serializations to represent sample inputs: for
functions, it is args.txt, for proc it is some other ad-hoc serialization,
and internal to the fuzzer datastructure it is serialized as
testvector::SampleInputsProto.

The tools have been prepared to take a single protobuffer. This CL
writes the protobuffer (as testvector.pbtxt) but still alongside
the 'old' serialization formats.

Part of the refactoring to universally use testvector::SampleInputsProto
(follow-up steps: use this format exclusivey and remove old
serialization formats).

Issues: https://github.com/google/xls/issues/1645
PiperOrigin-RevId: 694671946