Add some basic elements of type_system_v2.
This includes the inference table and its final conversion to type info, but with only some of the logic that will be needed in both (and only some of the eventual interface of the inference table).

Completely absent so far is the version of `TypecheckModule()` that will populate the inference table.

PiperOrigin-RevId: 696661361