Simplify Sample to just keep testvector proto.
This allows to add holdoff-data to the crasher input and
get it all the way to arriave at simulate_module_main.
A sample is included, which is just a copy of an existing
crasher with some valid-holdoffs added.

Make xls::Sample's primary constructor accept a testvector, but
also provide access to legacy args/channel_names data, mostly
used in unit tests now.

The separation of args() and channel_names() was confusing and
with holdoff data, there would be another piece of data that had
to be round-tripped to proto buffers and back. So this needs
to be one data structure, ideally directly backed by a proto
buffer.
For now, this will be the testvector proto (which then also allows
to be directly serialized and is already used in the reproducible
crasher files), but will probably need to be refined with better
serialization of Values (that are right now semicolon-separated
text representations) in follow-up changes.

Left to do: get the fuzzer to actually emit this holdoff data.

Issues: https://github.com/google/xls/issues/1645
PiperOrigin-RevId: 696688803