Fix crash with invocation inside type alias used in unroll_for!.
The issue was that `ContainedWithinFunction` was running on the `clog2` invocation in the type alias in the example code, in the context of deducing the unroll_for! body. There is an assumption that `ContainedWithinFunction` does not look at type definitions outside functions, because they would have been deduced before getting into a function that refers to them. However, the default behavior of `CloneAst` is to clone type definitions referred to within the cloned subtree, which triggers unwanted re-analysis.

So this change applies both the index replacer and the `PreserveTypeDefinitionsReplacer` when cloning the body.

https://github.com/google/xls/issues/1717

Fixes #1717

PiperOrigin-RevId: 696893384