[xls][mlir] Fix bug in array-to-bits, add support for ArrayConcatOp and speed up
We weren't declaring ArrayIndexOp and friends illegal, so the rewriter sometimes would not apply our rewrite patterns.

Also noticed that ArrayConcatOp was missing.

Also noticed it was super slow on a large module due to no parallelization, so added that.

Then noticed that XLS verification was super slow (dominant in multithreading as we verify per thread),
so fixed that.

PiperOrigin-RevId: 697543700