[DSLX:LS] Fix/tighten up URI/path conversions.
We now have a wrapper type to indicate when things have been checked to be in
URI form, and that type offers convenience APIs for conversion to `string_view`
for the underlying and to `std::filesystem::path` to strip off the URI prefix.

We also add a hook to the virtual filesystem layer to allow the LSP filesystem
implementation to validate that the "paths" being resolved are in fact all in
canonical URI form.

Previously we ran into trouble because the stdlib path was a relative path and
pointed directly at the filesystem, and thus was clashing with the expectation
that everything would be in URI form; i.e. a module would import a stdlib
module and then that would have a filesystem (non-URI) path, and that would
potentially import other stdlib modules with more filesystem (non-URI) paths.

We check as invariant in the file table that no URIs are held there -- the
conversion into and out of URI form happens on the import boundary.

This PR also plumbs virtualizable filesystem into the error printing routines
so they don't try to look at the real filesystem when they should be looking at
the virtualized filesystem contents, e.g. in any error reporting done for parse
errors encounterd by the LSP. Now that the LSP-parsed span filenames look
indistinguishable from real filesystem filenames we have to be a bit more
careful to use the virtual filesystem abstraction instead of the real
filesystem in retrieving file contents.

This also plumbs the virtual filesystem in everywhere to places that
GetContents is called to ensure consistency across the DSLX facilities. In
general we shouldn't need to rely on `//xls/common/file:filesystem` as a dep
except in tools that write output files (as the VirtualizableFileSystem doesn't
have SetContents).