[XLS][MLIR] Make `otherwise` parameter of `xls.sel` optional.
The XLS operation `sel` allows the `default` argument if and only if not all cases are specified, i.e., if, for an `n`-bit input, there are fewer than `2^n` cases provided. In particular, a `sel` operation cannot be constructed in XLS proper if all cases are provided *and* there is a `default` argument. This made translation to XLS (proper) fail if the MLIR op `sel` had all cases defined because the `otherwise` (corresponding to `default`) was mandatory.

This CL makes the `otherwise` parameter of `xls.sel` optional. As the op already has a variadic parameter, this requires making the op an `AttrSizedOperandSegments`. In order to avoid the verbosity of the generic op form that this entails, the CL also adds a custom assembly format that omits the `otherwise` parameter if unspecified and adapts to LIT tests that use that op.

PiperOrigin-RevId: 700418691