[xls][mlir] Sort spawns regions topologically after OptimizeSpawns finishes
We apply patterns that modify two regions, the next and spawns regions. The edits
to the next region always obey dominance rules, but the changes to the spawns regions
are not guaranteed to (we don't have a perfect place to insert our new spawn instruction).

Rather than trying to guarantee it which may be hard, simply sort the spawns regions after
the optimization runs.

PiperOrigin-RevId: 701078946