[DSLX:cleanup] Remove ConstantArray node from AST, it is misleading / not useful.
Previously ConstantArray was trying to convey some type-system-level properties
as imbued in an AST node, but it was not fully accurate, and generally only
served to confuse things. This gets us back to where the AST just represents
the parsed syntax and the typechecker figures out constexpr properties and
annotates them in the type information. This also allows us to seal up the
Array AST node as final  — subclassing of AST nodes can be awkward, we
generally use hierarchy in the AST very sparingly.

This reworks the deduce routine a bit to consolidate on the main discrepancy
that was there between Arrays and ConstArrays, which was that we would, for
user convenience, annotate the element type on any un-annotated numerical
literals contained within the array. This will become better and more general
will type inference 2.0 but until that comes we want to make sure we don’t have
to change code in a sprawling way.

Just trying to tighten up invariants in advance of somebody investigating
google/xls#1748