[xls] Performance improvements for BinaryDecisionDiagram::GetOrCreateNode
BinaryDecisionDiagram::GetOrCreateNode() was observed to be extremely hot on several profiles of real workloads.

In the case where the BddFunction has large bits[N] types, we end up calling BinaryDecisionDiagram::NewVariable()
N times per xls::Node. The decision diagram's nodes_ can be dominated by "base" variable nodes, especially when the
work that the BDD does is sparse (non-trivial expressions are created for a small subset of the node count).

We optimize in two ways:
  * "bare" variable nodes get allocated in their own vector, not on the nodes_map_. The nodes_map_ was extremely
      large and nodes_map_.find() during GetOrCreateNode() was dominating (note that find on a large map where
      the searched-for item does not exist in the map is the worst case runtime).
  * A bulk API "NewVariables(N)" is created. The implementation is a little tricky to try to mitigate any
      performance drop if N is small (the naive "reserve on all vectors" approach would result in linear numbers
      of vector reallocations as it elides push_back's exponential resize optimization).

PiperOrigin-RevId: 702027421