Implement direct-ins to sub-blocks.
Receives from single value channels are meant to be realizable in hardware as
wires. There isn't a mechanism in codegen to support multiple receives
in different procs/blocks on the same single value channel while retaining this
property and avoiding registering of the values.

The approach used here is to create an internal single value channel
for each sub-block through which the value(s) from the external single value channel(s) are forwarded, with the intention of being able to end up with wires
only and no registering of the direct-in / single-value inputs.

PiperOrigin-RevId: 704379181