[xls][mlir] Add next_value for eproc.
The XLS proc, in contrast with DSLX proc, supports multiple next_value instances. Enable this by adding new op that is used to represent a predicated merge of next_values. The op, similar to XLS, assumes that the predicates are disjoint at runtime. This op has no semantics outside its usage as operand for eproc's yield.

This does not yet a better custom syntax, but it would be a welcome addition :)

PiperOrigin-RevId: 705139951