Simplify the declaration type checking for tuple variables in type_system_v2.
The strategy of trying to apply a type annotation from the LHS directly to the RHS, in a case like `const X: (u32, u32) = (1, 2)` was less clean than relying on unification of the 2 inferred tuple types at the end, and it would not work as well for arrays.

The downside to just ripping out the extra logic in typecheck_module_v2.cc would have been opaque error messages upon failure, like "type mismatch: (TypeVariableTypeAnnotation, TypeVariableTypeAnnotation) vs. (uN[1], uN[2])". So, this change also adds deep resolution of TypeVariableTypeAnnotation.

This change also corrects the signedness and bit count unification rules, where the logic was not quite matching the original intent.

PiperOrigin-RevId: 705488960