[xls][mlir] Fix extractAsTopModule(SprocOp)
When checking which external callables should be cloned, we were only visiting
the getCallableRegion(). There can only be one callable region, so we pick
"spawns" which is where outgoing calls to other SprocOps occur. But we can also
call functions in "next", so simply check all regions of the CallableOpInterface.

PiperOrigin-RevId: 706635915