[xlscc] Add option to set control channel FIFO depth for multi-block designs.
Sets a default FIFO depth of 0 for sub-block control channels,
and adds the [[hls_control_channel_depth(N)]] Clang attribute to allow
the depth to be explicitly set.

xlscc represents C++ control flow by putting IO operations
in a serial token chain according to the order in the code.
This also applies to IO operations for calls to sub-blocks,
specifically in this case the send on the control channel.

PiperOrigin-RevId: 706841832