Allow configuration of i/o flop on per-channel basis in IR
This adds support for input_flop_kind and output_flop_kind on channels. If present these override the codegen `--flop_inputs[_kind=...]` and `--flop_outputs[_kind=...]` flags for ports associated with those channels. These should be generated by the frontend or added by some other pass. These options have no effect on the semantics of the proc ir and are only a directive for codegen conversion.

PiperOrigin-RevId: 706887241