Enable references to global constants in parametric defaults.
This basically requires more careful control of the order of conversion from InferenceTable to TypeInfo, because ConstexprEvaluator presumes all prerequisites are proactively loaded.

This also changes inference_table_test to be more of a pure unit test for the table functions, because we are now at the point where equivalent end-to-end testing is being done with less hackery in typecheck_module_v2_test. The hackery previously in inference_table_test relied on the table being the source of node order.

PiperOrigin-RevId: 709194186