Add a utility to "zip" 2 AST subtrees.
This is similar in concept to `ZipTypes`, and will also be used in type inference, but in type_system_v2 it is generally annotations rather than Type objects that we would need to zip. And there will be broader uses for it than in V1, such as pattern matching a concrete type annotation to a parameterized one to infer the parametrics.

PiperOrigin-RevId: 710717388