Add a utility to "solve for" the parametrics in a type annotation.
This is basically the type_system_v2 counterpart of ParametricBind(), which decides what `N` is in a situation like

fn foo<N: u32>(a: uN[N]) { ... }
foo(u24:5);

where the non-parametric actual argument is all it has to go by.

v1 does this against `Type` objects that are not fully concretized. v2 does it against `TypeAnnotation` objects, because in v2 you only have a `Type` when you are ready for full concretization.

PiperOrigin-RevId: 710727872