[XLS] Remove conditional specialization for operands of Boolean operations
It's not safe to specialize all operands of a Boolean operation at the same time based on visibility; this can only be done asymmetrically, or else we risk accidentally excluding cases where all operands are true.

For example, it's true that when considering

    OR(A, B),

we can assume that B is false while simplifying A, since the value of A is irrelevant whenever B is true. We can similarly assume that A is false while simplifying B.

However... we can't do these at the same time. To see why, suppose it turns out that A = X and B = X (i.e., that A and B are logically equivalent). In this case, we might try to simplify

    OR(A, B)

by replacing A with its value whenever B = 0 (i.e., replacing A with 0) and B with its value whenever A = 0 (i.e., replacing B with 0). This means we'd replace the expression with:

    OR(0, 0) = 0.

However, if X is true, OR(A, B) = OR(1, 1) = 1.

PiperOrigin-RevId: 710762944