[codegen] Add option to shuffle Verilog output order
As a way to exercise the rest of a synthesis pipeline, we add a command-line flag that will shuffle the order of the Verilog output from codegen with a seeded pseudorandom shuffler. We keep the order similar - register declarations, then logic, then register assignments - and respect topological ordering within the logic section.

PiperOrigin-RevId: 711779296