[xls][mlir] Add LiteralOp to XLS MLIR Dialect.
This operation constructs a literal of arbitrarily complex type using a
body region that construct this value using basic scalar operations.

It more closely resembles the XLS literal IR node for loss-less
conversion between the IR and MLIR, but is not intended to be used
during complex operations and transforms in MLIR due to the lack of
constant folding.

A pass to remove this operation by inlining (and potentially a pass to
reconstruct such literal ops by grouping ops) will be added in the
future.

Only a small subset of XLS operations is permitted inside the body
region. The new LiteralMemberOpInterface is used to identify such
ops.