Unify the unification of signedness and size.
The prior revision of this logic had an incorrect bias toward adjusting the signedness of auto annotations only if they appeared to the right of explicit ones. This change makes the signedness handling as insensitive as the bit count handling to which side the auto term is on. The two pieces of logic are unified here due to the possible effect of signedness coercion on bit count.

PiperOrigin-RevId: 713656605