[XLS] Conditional specialization: allow recursive analysis through NOT
We accidentally missed recursion through NOT in the previous change that added support for propagating conditions through logic.

With this fixed, we find that in some cases, we can end up recomputing the same extremely deep recursions repeatedly. To avoid this, we introduce a cache of the ConditionSets implied by a given Condition.

PiperOrigin-RevId: 715886243