Add `TypeAnnotation` subclasses representing the idea of decltype(container.member).
Change TypecheckModuleV2 to use these instead of the direct member/element type annotations.

For the scenarios handled at head, this is ultimately a no-op change, but describing the types of parametric struct members (or their contents, when they are themselves containers) will require it to be done this way in order to avoid leakage of parametrics.

PiperOrigin-RevId: 718553894