[DSLX:TIv2] Implement cast operation.
The type annotation the user provides is the type annotation that results from
the cast expression.

It is validated when we go to finalize the inference table into type
information, where we have concrete types that can check whether it is an
acceptable cast.

Note that it'd be nice in the future to not need to annotate types when the type is available from the cast operation, e.g.

```dslx
const FOO = u32:42;
const BAR = FOO as s32;  // s32 is annotated here, why put it on the const as well?
```

but I couldn't figure out how to do that in this PR.

Note that I also switched from S64 to U32s for default "usize like" usage and
put a link to the relevant github issue where it explains why it'd be nice to
have a usize type for this purpose similar to Rust.