[XLS] Fix proc-state-legalization for state-reads that influence next_values
If a next-value (state write)'s predicate depends on the read of that same state element, we would previously add the predicate to the state read - introducing a cycle in the graph, and causing codegen to fail with an error.

To fix this, we automatically use a variant of the write's predicate, removing all dependency on the state read's value, but guaranteeing that the new predicate is true whenever the original expression would have been true.

PiperOrigin-RevId: 719457555