Have inlining pass append invoke's locations to the inlined nodes.
When we inline a function information about where in the source the invoke occurred is lost. To preserve it we can just add it onto the end of any nodes which are 'inlined'. This means this information is still around for later debugging and allows one to distinguish between two different invokes nodes.

PiperOrigin-RevId: 720327368