Recognize bit-slice based comparisons in back-propagation
In some circumstances we will implement unsigned comparisons against powers of 2 as a check that the high bits of the variable are all zero (or is non-zero as appropriate).

For example

```
  ult(X, 4)
```

can be written as

```
  eq(bitslice(X, 2, width(X) - 2), 0)
```

The back propagation needs to specifically handle these cases to ensure that best results can be achieved. Without this we missed some simple cases of proc state narrowing.

PiperOrigin-RevId: 721036997