Add tool to extract state update logic from a proc.
Sometimes when investigating optimizations its useful to cut down procs to only certain state elements. This performs that minimization automatically creating a proc where only the state elements requested and the logic needed to update them is included. Other state elements are replaced with 'recv's to get their values. All sends are removed as are any recvs that do not directly contribute to the value of the state elements. Optionally a channel which sends the value of the state element on each activation is inserted.

PiperOrigin-RevId: 721049820