Defer resolution of invoked functions in type_system_v2.
Previously, we would determine the callee function on encountering an invocation node, which is an easy way to handle static function references, but doesn't work well for impls or function objects that can be aliased.

With this change, we will
- "Typecheck" (i.e. populate the InferenceTable) for the whole module, including parametric functions, but not apply formal argument types to actual types.
- Resolve the function being invoked by each invocation at conversion time, apply the formal types to the actual types, and generate invocation-specific type info as needed.

This also allows us to get rid of the complexity of "invocation scoped" type annotations, in favor of cleansing those annotations of their parametrics, as we do with struct members. The removal of these annotations also simplifies the conversion order determination.

PiperOrigin-RevId: 721470940