[xls][mlir] Fix foreign function calling where the function returns a tuple
Somewhere in its codegen, XLS collapses a tuple result of a top function to
its bits form. Therefore the generated verilog has only a single result, not
three, for a floating point value.

However the ffi_proto annotation is typechecked that it has the correct number
of results in it for the type of the call... which forces us to write an invalid
ffi_proto.

Work around this by:
  1) Modifying the imported function's result type so it returns the concat of the
       tuple elements.
  2) Re-tuplifying when necessary after we generate an Invoke so that the translation
       type checks.

Related https://github.com/google/xls/issues/1722

PiperOrigin-RevId: 721656919