[xls] Work hard to ensure few symbol conflicts
We had many symbols having the same stem, which were uniquified by the SymbolTable.

But this not only made symbols with names like foo_40294, but was subtly nondeterministic.

I haven't yet found the cause of the nondeterminism, but given that there should be few
auto-uniquifications now, it should be easier to track down if it still exists.

PiperOrigin-RevId: 721681462