Fix bug with arg and fn TypeAnnotation storage for parametric functions.
An example scenario which fails before this change is:

```
fn bar<A: u32>(a: uN[A]) -> uN[A] { a + 1 }
fn foo<A: u32>(a: uN[A]) -> uN[A] { bar<A>(a) }
const X = foo<24>(4);
const Y = foo<32>(5);
```

What we would do is attach type annotations to the "bar" NameRef in line 2 and the "a" NameRef in line 2 that would only be valid in the context of the line 3 invocation, and would make line 4 fail unification. This was true even before the recent re-factor.

The current change adds the notion of type annotations that are only associated with a variable within a specific parametric invocation, to resolve the issue.

PiperOrigin-RevId: 721813931