[XLS] Fix bug in BddQueryEngine::SpecializeGiven
Previously, when told to specialize on an interval set, if the check for one component interval was saturated, it would be omitted from the final assumed disjunction... making the assumption *stronger* than intended.

This could make values appear to be determined when they were not.

Found by fuzzing a future change.

PiperOrigin-RevId: 723593153