Rename ParametricInvocation to ParametricContext and make it usable for structs.
So far we have gotten away with just erasing parametrics from struct static field types, but when it comes to parametric impls, we need a context against which to analyze nontrivial constant declarations.

The actual use of contexts for impl constants that will be in a follow-up. This is just a no-op rename/refactor.

PiperOrigin-RevId: 723961726