[XLS] Fix conditional-specialization's handling of no-op `next_value` nodes
I'd previously introduced logic that ignored these nodes; however, this could create mis-compilations in rare circumstances - where this resulted in specializations that changed a node that fed a no-op `next_value` predicate - which mostly only seem to occur with a forthcoming change in place. We fix this by instead allowing conditional specialization (when invoked before proc-state legalization) to remove these nodes before analyzing the code.

In exchange, we remove the (now obsolete) code that was meant to fix this issue downstream in proc-state legalization... and update a couple of other passes to properly handle more situations involving `next_value` nodes.

PiperOrigin-RevId: 723962607